todcore

+++ todcore/include/tod/platformdef.h
@@ -10,8 +10,8 @@ typedef char int8;
 typedef unsigned char uint8;
 typedef short int16;
 typedef unsigned short uint16;
-typedef long int32;
-typedef unsigned long uint32;
+typedef int int32;
+typedef unsigned int uint32;
 
 #ifdef PLATFORM_MAC
 #include <cxxabi.h>
